+++ LKW/.gitignore
@@ -1,3 +1,3 @@
 .idea
-.out
+out
 *.iml
