--- LKW/baekjoon/22_03_02/SugarDelivery.java
@@ -1,20 +1,0 @@
-import java.util.Scanner;
-
-public class Main {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        int num = sc.nextInt();
-        int rm = -1;
-        Loop1:
-        for(int tr=0;tr*3<=num;tr++){
-            Loop2:
-            for(int fv = num/5;fv>=0;fv--){
-                if (num == fv*5+tr*3){
-                    rm =fv + tr;
-                    break Loop1;
-                }
-            }
-        }
-        System.out.println(rm);
-    }
-}
